OmeroBuild : Reworking Ivy while trying to fix jbossall-client
jbossall-client contains hibernate-search classes which are out
of sync with our hibernate-search version, causing invalid class
exceptions and similar.

In an attempt to fix this, ivy was upgraded to 2.0.0-beta2 as well
as having a new top-level build.xml written. This ultimately did not work.

The only solution for the moment is to ship our own jbossall-client
jar with the hibernate-search classes removed. Hopefully,
the next upgrade of jboss will correct this.


git-svn-id: file:///home/svn/omero/trunk@2542 05709c45-44f0-0310-885b-81a1db45b4a6