Review all instances of whitelist/blacklist
This does _not_ try to maintain backwards compatibility
since the related methods and classes are internal. The
likeliest impact is that of the change in property name.
A redirecting property could be added here, but a nicer
choice is likely to automatically upgrade in omero-py.

see: https://github.com/ome/bioformats/issues/3580